requirements.txt
awesome-slugify
beautifulsoup4